+++ src/main/java/com/phamxuantoan/webshop/repository/CatalogRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CatalogRepository extends JpaRepository<CatalogEntity, Integer> {
+public interface CatalogRepository extends JpaRepository<CatalogEntity, Long> {
     Catalog findByCatalogName(String catalogName);
 }
