+++ src/main/java/com/phamxuantoan/webshop/repository/CommentRepository.java
@@ -10,11 +10,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CommentRepository extends JpaRepository<CommentEntity, Integer> {
+public interface CommentRepository extends JpaRepository<CommentEntity, Long> {
 
     @Query(value ="select * from comment  where comment.user_id = ?1", nativeQuery = true)
-    public List<CommentEntity> findCommentByUserID(Integer uid);
+    public List<CommentEntity> findCommentByUserID(Long uid);
 
     @Query(value ="select * from comment  where comment.product_id = ?1", nativeQuery = true)
-    public List<CommentEntity> findCommentByProductID(Integer uid);
+    public List<CommentEntity> findCommentByProductID(Long uid);
 }
