+++ src/main/java/com/phamxuantoan/webshop/repository/OrderDetailRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderDetailRepository extends JpaRepository<OrderDetailEntity, Integer> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetailEntity, Long> {
     @Query(value = "select * from billsdetail  where billsdetail.bills_id = ?1",nativeQuery = true)
-    List<OrderDetailEntity> findBillByDetailId(Integer id);
+    List<OrderDetailEntity> findBillByDetailId(Long id);
 }
