+++ src/main/java/com/phamxuantoan/webshop/service/IOrderDetailService.java
@@ -7,8 +7,8 @@ import java.util.List;
 
 public interface IOrderDetailService {
     List<OrderDetailDTO> findAll();
-    OrderDetailDTO getBillDetailById(Integer id);
-    List<OrderDetailDTO> getDetailByBillId(Integer id);
+    OrderDetailDTO getBillDetailById(Long id);
+    List<OrderDetailDTO> getDetailByBillId(Long id);
     OrderDetailDTO saveOrUpdate(OrderDetailEntity billsdetail);
-    void delete(Integer[] ids);
+    void delete(Long[] ids);
 }
